requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
ironbook-sdk>=0.3.1
ironbook-a2a-extension>=0.0.2
pydantic>=2.7.0
python-dotenv>=1.0.1
